Move ChromeToPaintMsg to gfx_traits.
This allows compositing not to depend on gfx.
