+++ components/compositing/Cargo.toml
@@ -9,7 +9,6 @@ name = "compositing"
 path = "lib.rs"
 
 [dependencies]
-gfx = {path = "../gfx"}
 gfx_traits = {path = "../gfx_traits"}
 script_traits = {path = "../script_traits"}
 style_traits = {path = "../style_traits"}
