+++ components/compositing/compositor.rs
@@ -14,8 +14,7 @@ use euclid::rect::TypedRect;
 use euclid::scale_factor::ScaleFactor;
 use euclid::size::TypedSize2D;
 use euclid::{Matrix4D, Point2D, Rect, Size2D};
-use gfx::paint_thread::{ChromeToPaintMsg, PaintRequest};
-use gfx_traits::{ScrollPolicy, StackingContextId};
+use gfx_traits::{ChromeToPaintMsg, PaintRequest, ScrollPolicy, StackingContextId};
 use gfx_traits::{color, Epoch, FrameTreeId, FragmentType, LayerId, LayerKind, LayerProperties};
 use gleam::gl;
 use gleam::gl::types::{GLint, GLsizei};
