+++ components/compositing/lib.rs
@@ -15,7 +15,6 @@ extern crate app_units;
 
 extern crate azure;
 extern crate euclid;
-extern crate gfx;
 extern crate gfx_traits;
 extern crate gleam;
 extern crate image;
@@ -39,7 +38,7 @@ extern crate webrender_traits;
 pub use compositor_thread::CompositorProxy;
 pub use compositor::IOCompositor;
 use euclid::size::TypedSize2D;
-use gfx::paint_thread::ChromeToPaintMsg;
+use gfx_traits::ChromeToPaintMsg;
 use ipc_channel::ipc::IpcSender;
 use msg::constellation_msg::PipelineId;
 use script_traits::{ConstellationControlMsg, LayoutControlMsg};
